Move the mirrorer to a separate directory
