+++ ydb/core/persqueue/pqtablet/partition/partition.cpp
@@ -1,4 +1,3 @@
-#include "mirrorer_factory.h"
 #include "offload_actor.h"
 #include "partition_common.h"
 #include "partition_compactification.h"
@@ -14,6 +13,7 @@
 #include <ydb/core/persqueue/pqtablet/common/event_helpers.h>
 #include <ydb/core/persqueue/pqtablet/common/logging.h>
 #include <ydb/core/persqueue/pqtablet/common/tracing_support.h>
+#include <ydb/core/persqueue/pqtablet/partition/mirrorer/mirrorer_factory.h>
 #include <ydb/core/persqueue/writer/source_id_encoding.h>
 #include <ydb/core/protos/counters_pq.pb.h>
 #include <ydb/core/protos/msgbus.pb.h>
